+++ server/package.json
@@ -7,8 +7,8 @@
   "scripts": {
     "start": "node build/server.js",
     "serve": "node --loader ts-node/esm src/server.ts",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "tsc"
+    "build": "tsc",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js"
   },
   "author": "",
   "license": "ISC",
@@ -20,9 +20,38 @@
     "winston": "^3.6.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.17.4",
+    "@babel/preset-typescript": "^7.16.7",
     "@types/express": "^4.17.13",
+    "@types/jest": "^27.4.0",
     "@types/node-fetch": "^3.0.3",
+    "@types/supertest": "^2.0.11",
+    "babel-plugin-replace-import-extension": "^1.1.2",
+    "jest": "^27.5.1",
+    "supertest": "^6.2.2",
+    "ts-jest": "^27.1.3",
     "ts-node": "^10.5.0",
     "typescript": "^4.5.5"
+  },
+  "jest": {
+    "preset": "ts-jest/presets/default-esm",
+    "globals": {
+      "ts-jest": {
+        "useESM": true,
+        "diagnostics": {
+          "ignoreCodes": ["TS2691"]
+        }
+      }
+    },
+    "moduleFileExtensions": [
+      "js",
+      "ts"
+    ],
+    "extensionsToTreatAsEsm": [
+      ".ts"
+    ],
+    "moduleNameMapper": {
+      "^(\\.{1,2}/.*)\\.js$": "$1"
+    }
   }
 }
